src/orders/test/integ/requirements.txt
pytest==7.1.1
python-dotenv
../../testing